Repair ZIP Code or ZCTA Vector

This function repairs two of the four conditions identified in the validation checks with zi_validate(). For the other two conditions, values are conveted NA. See Details below for the specific changes made.

zi_repair(x, style = "zcta5")

Arguments

x

A vector containing ZIP or ZCTA values to be repaired.

style

A character scalar - either "zcta5" or "zcta3".

Value

A repaired vector of ZIP or ZCTA values.

Details

The zi_repair() function addresses four conditions:

If the input vector is numeric, it will be converted to character data.

If there are values less than five characters (if style = "zcta5", the default), or three characters (if style = "zcta3"), they will be padded with leading zeros.

If there are input values over five characters (if style = "zcta5", the default), or three characters (if style = "zcta3"), they will be converted to NA.

If there are input values that have non-numeric characters, they will be converted to NA.

Since two of the four steps will result in NA values, it is strongly recommended to attempt to manually fix these issues first.

Examples

# sample five-digit ZIPs with character
zips <- c("63088", "63108", "zip")

# failed validation
zi_validate(zips)
#> [1] FALSE

# repair
zips <- zi_repair(zips)
#> Warning: NAs introduced by coercion

# successful validation
zi_validate(zips)
#> [1] TRUE