Download Demographic Data for Five-digit ZCTAs

This function returns demographic data for five-digit ZIP Code Tabulation Areas (ZCTAs), which are rough approximations of many (but not all) USPS ZIP codes.

Usage

zi_get_demographics(year, variables = NULL, table = NULL,
    survey, output = "tidy", zcta = NULL, key = NULL)

Arguments

year

A four-digit numeric scalar for year. zippeR currently supports data for from 2010 to 2022. Different survey products are available for different years. See the survey parameter for more details

variables

A character scalar or vector of variable IDs.

table

A character scalar of a table ID (only one table may be requested per call).

survey

A character scalar representing the Census product. It can be either a Decennial Census product (either "sf1" or "sf3") or an American Community Survey product (either "acs1", "acs3", or "acs5"). For Decennial Census calls, only the 2010 Census is available. In addition, if a variable cannot be found in "sf1", the function will look in "sf3". Also note that "acs3" was discontinued after 2013.

output

A character scalar; one of "tidy" (long output) or "wide" depending on the type of data format you want. If you are planning to pass these data to zi_aggregate(), you must choose "tidy". If you are leaving these data as five-digit ZCTAs and are planning to join them with geometric data, "wide" is the strongly encouraged format.

zcta

An optional vector of ZCTAs that demographic data are requested for. If this is NULL, data will be returned for all ZCTAs. If a vector is supplied, only data for those requested ZCTAs will be returned. The vector can be created with zi_get_geometry() and should only contain five-digit ZCTAs.

key

A Census API key, which can be obtained at https://api.census.gov/data/key_signup.html. This can be omitted if tidycensus::census_api_key() has been used to write your key to your .Renviron file. You can check whether an API key has been written to .Renviron by using Sys.getenv("CENSUS_API_KEY").

Value

A tibble containing all demographic data requested in either "tidy" or "wide" format.

Examples

# \donttest{
  # download all ZCTAs
  zi_get_demographics(year = 2012, variables = "B01003_001", survey = "acs5")
#> # A tibble: 33,120 × 4
#>    GEOID variable   estimate   moe
#>    <chr> <chr>         <dbl> <dbl>
#>  1 00601 B01003_001    18544   271
#>  2 00602 B01003_001    41640   147
#>  3 00603 B01003_001    54540   735
#>  4 00606 B01003_001     6593   270
#>  5 00610 B01003_001    29141   180
#>  6 00612 B01003_001    69017  1500
#>  7 00616 B01003_001    10750  1084
#>  8 00617 B01003_001    24733   171
#>  9 00622 B01003_001     5923   984
#> 10 00623 B01003_001    44943   984
#> # ℹ 33,110 more rows

  # limit output to subset of ZCTAs
  ## download all ZCTAs in Missouri, intersects method
  mo20 <- zi_get_geometry(year = 2020, state = "MO", method = "intersect")
#> ZCTAs can take several minutes to download.  To cache the data and avoid re-downloading in future R sessions, set `options(tigris_use_cache = TRUE)`
#> Downloading: 80 kB     Downloading: 80 kB     Downloading: 80 kB     Downloading: 80 kB     Downloading: 80 kB     Downloading: 80 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 370 kB     Downloading: 370 kB     Downloading: 370 kB     Downloading: 370 kB     Downloading: 370 kB     Downloading: 370 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 830 kB     Downloading: 830 kB     Downloading: 850 kB     Downloading: 850 kB     Downloading: 850 kB     Downloading: 850 kB     Downloading: 870 kB     Downloading: 870 kB     Downloading: 950 kB     Downloading: 950 kB     Downloading: 950 kB     Downloading: 950 kB     Downloading: 950 kB     Downloading: 950 kB     Downloading: 1,000 kB     Downloading: 1,000 kB     Downloading: 1,000 kB     Downloading: 1,000 kB     Downloading: 1,000 kB     Downloading: 1,000 kB     Downloading: 1.2 MB     Downloading: 1.2 MB     Downloading: 1.2 MB     Downloading: 1.2 MB     Downloading: 1.3 MB     Downloading: 1.3 MB     Downloading: 1.3 MB     Downloading: 1.3 MB     Downloading: 1.8 MB     Downloading: 1.8 MB     Downloading: 1.8 MB     Downloading: 1.8 MB     Downloading: 1.8 MB     Downloading: 1.8 MB     Downloading: 1.8 MB     Downloading: 1.8 MB     Downloading: 1.9 MB     Downloading: 1.9 MB     Downloading: 1.9 MB     Downloading: 1.9 MB     Downloading: 1.9 MB     Downloading: 1.9 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2.1 MB     Downloading: 2.1 MB     Downloading: 2.1 MB     Downloading: 2.1 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.7 MB     Downloading: 2.7 MB     Downloading: 2.7 MB     Downloading: 2.7 MB     Downloading: 3.4 MB     Downloading: 3.4 MB     Downloading: 3.4 MB     Downloading: 3.4 MB     Downloading: 4 MB     Downloading: 4 MB     Downloading: 4 MB     Downloading: 4 MB     Downloading: 4 MB     Downloading: 4 MB     Downloading: 4.2 MB     Downloading: 4.2 MB     Downloading: 4.2 MB     Downloading: 4.2 MB     Downloading: 4.2 MB     Downloading: 4.2 MB     Downloading: 5.2 MB     Downloading: 5.2 MB     Downloading: 5.2 MB     Downloading: 5.2 MB     Downloading: 5.2 MB     Downloading: 5.2 MB     Downloading: 6.7 MB     Downloading: 6.7 MB     Downloading: 6.7 MB     Downloading: 6.7 MB     Downloading: 6.7 MB     Downloading: 6.7 MB     Downloading: 8.3 MB     Downloading: 8.3 MB     Downloading: 8.8 MB     Downloading: 8.8 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 15 MB     Downloading: 15 MB     Downloading: 17 MB     Downloading: 17 MB     Downloading: 19 MB     Downloading: 19 MB     Downloading: 20 MB     Downloading: 20 MB     Downloading: 22 MB     Downloading: 22 MB     Downloading: 24 MB     Downloading: 24 MB     Downloading: 25 MB     Downloading: 25 MB     Downloading: 27 MB     Downloading: 27 MB     Downloading: 27 MB     Downloading: 27 MB     Downloading: 27 MB     Downloading: 27 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 28 MB     Downloading: 29 MB     Downloading: 29 MB     Downloading: 29 MB     Downloading: 29 MB     Downloading: 29 MB     Downloading: 29 MB     Downloading: 30 MB     Downloading: 30 MB     Downloading: 32 MB     Downloading: 32 MB     Downloading: 34 MB     Downloading: 34 MB     Downloading: 35 MB     Downloading: 35 MB     Downloading: 37 MB     Downloading: 37 MB     Downloading: 39 MB     Downloading: 39 MB     Downloading: 40 MB     Downloading: 40 MB     Downloading: 42 MB     Downloading: 42 MB     Downloading: 44 MB     Downloading: 44 MB     Downloading: 45 MB     Downloading: 45 MB     Downloading: 47 MB     Downloading: 47 MB     Downloading: 48 MB     Downloading: 48 MB     Downloading: 48 MB     Downloading: 48 MB     Downloading: 48 MB     Downloading: 48 MB     Downloading: 48 MB     Downloading: 48 MB     Downloading: 49 MB     Downloading: 49 MB     Downloading: 51 MB     Downloading: 51 MB     Downloading: 53 MB     Downloading: 53 MB     Downloading: 54 MB     Downloading: 54 MB     Downloading: 56 MB     Downloading: 56 MB     Downloading: 58 MB     Downloading: 58 MB     Downloading: 59 MB     Downloading: 59 MB     Downloading: 61 MB     Downloading: 61 MB     Downloading: 63 MB     Downloading: 63 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 64 MB     Downloading: 65 MB     Downloading: 65 MB     Downloading: 65 MB     Downloading: 65 MB     Downloading: 65 MB     Downloading: 65 MB     Downloading: 67 MB     Downloading: 67 MB     Downloading: 68 MB     Downloading: 68 MB     Downloading: 68 MB     Downloading: 68 MB     Downloading: 68 MB     Downloading: 68 MB     Downloading: 69 MB     Downloading: 69 MB     Downloading: 69 MB     Downloading: 69 MB     Downloading: 69 MB     Downloading: 69 MB     Downloading: 70 MB     Downloading: 70 MB     Downloading: 72 MB     Downloading: 72 MB     Downloading: 72 MB     Downloading: 72 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 74 MB     Downloading: 75 MB     Downloading: 75 MB     Downloading: 75 MB     Downloading: 75 MB     Downloading: 75 MB     Downloading: 75 MB     Downloading: 76 MB     Downloading: 76 MB     Downloading: 76 MB     Downloading: 76 MB     Downloading: 76 MB     Downloading: 76 MB     Downloading: 76 MB     Downloading: 76 MB     Downloading: 77 MB     Downloading: 77 MB     Downloading: 77 MB     Downloading: 77 MB     Downloading: 77 MB     Downloading: 77 MB     Downloading: 77 MB     Downloading: 77 MB     Downloading: 77 MB     Downloading: 77 MB     Downloading: 79 MB     Downloading: 79 MB     Downloading: 79 MB     Downloading: 79 MB     Downloading: 80 MB     Downloading: 80 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 81 MB     Downloading: 82 MB     Downloading: 82 MB     Downloading: 82 MB     Downloading: 82 MB     Downloading: 82 MB     Downloading: 82 MB     Downloading: 82 MB     Downloading: 82 MB     Downloading: 82 MB     Downloading: 82 MB     Downloading: 83 MB     Downloading: 83 MB     Downloading: 83 MB     Downloading: 83 MB     Downloading: 83 MB     Downloading: 83 MB     Downloading: 83 MB     Downloading: 83 MB     Downloading: 83 MB     Downloading: 83 MB     Downloading: 84 MB     Downloading: 84 MB     Downloading: 84 MB     Downloading: 84 MB     Downloading: 84 MB     Downloading: 84 MB     Downloading: 86 MB     Downloading: 86 MB     Downloading: 86 MB     Downloading: 86 MB     Downloading: 86 MB     Downloading: 86 MB     Downloading: 86 MB     Downloading: 86 MB     Downloading: 86 MB     Downloading: 86 MB     Downloading: 88 MB     Downloading: 88 MB     Downloading: 89 MB     Downloading: 89 MB     Downloading: 91 MB     Downloading: 91 MB     Downloading: 93 MB     Downloading: 93 MB     Downloading: 94 MB     Downloading: 94 MB     Downloading: 95 MB     Downloading: 95 MB     Downloading: 95 MB     Downloading: 95 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 96 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 97 MB     Downloading: 98 MB     Downloading: 98 MB     Downloading: 98 MB     Downloading: 98 MB     Downloading: 98 MB     Downloading: 98 MB     Downloading: 98 MB     Downloading: 98 MB     Downloading: 98 MB     Downloading: 98 MB     Downloading: 98 MB     Downloading: 98 MB     Downloading: 99 MB     Downloading: 99 MB     Downloading: 99 MB     Downloading: 99 MB     Downloading: 99 MB     Downloading: 99 MB     Downloading: 99 MB     Downloading: 99 MB     Downloading: 99 MB     Downloading: 99 MB     Downloading: 99 MB     Downloading: 99 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 100 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 110 MB     Downloading: 120 MB     Downloading: 120 MB     Downloading: 120 MB     Downloading: 120 MB     Downloading: 120 MB     Downloading: 120 MB     Downloading: 120 MB     Downloading: 120 MB     Downloading: 120 MB     Downloading: 120 MB     Downloading: 120 MB     Downloading: 120 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 130 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 140 MB     Downloading: 150 MB     Downloading: 150 MB     Downloading: 150 MB     Downloading: 150 MB     Downloading: 150 MB     Downloading: 150 MB     Downloading: 150 MB     Downloading: 150 MB     Downloading: 150 MB     Downloading: 150 MB     Downloading: 150 MB     Downloading: 150 MB     Downloading: 160 MB     Downloading: 160 MB     Downloading: 160 MB     Downloading: 160 MB     Downloading: 160 MB     Downloading: 160 MB     Downloading: 160 MB     Downloading: 160 MB     Downloading: 160 MB     Downloading: 160 MB     Downloading: 160 MB     Downloading: 160 MB     Downloading: 170 MB     Downloading: 170 MB     Downloading: 170 MB     Downloading: 170 MB     Downloading: 170 MB     Downloading: 170 MB     Downloading: 170 MB     Downloading: 170 MB     Downloading: 170 MB     Downloading: 170 MB     Downloading: 170 MB     Downloading: 170 MB     Downloading: 180 MB     Downloading: 180 MB     Downloading: 180 MB     Downloading: 180 MB     Downloading: 180 MB     Downloading: 180 MB     Downloading: 180 MB     Downloading: 180 MB     Downloading: 180 MB     Downloading: 180 MB     Downloading: 180 MB     Downloading: 180 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 190 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 200 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 210 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 220 MB     Downloading: 230 MB     Downloading: 230 MB     Downloading: 230 MB     Downloading: 230 MB     Downloading: 230 MB     Downloading: 230 MB     Downloading: 230 MB     Downloading: 230 MB     Downloading: 230 MB     Downloading: 230 MB     Downloading: 230 MB     Downloading: 230 MB     Downloading: 240 MB     Downloading: 240 MB     Downloading: 240 MB     Downloading: 240 MB     Downloading: 240 MB     Downloading: 240 MB     Downloading: 240 MB     Downloading: 240 MB     Downloading: 240 MB     Downloading: 240 MB     Downloading: 240 MB     Downloading: 240 MB     Downloading: 250 MB     Downloading: 250 MB     Downloading: 250 MB     Downloading: 250 MB     Downloading: 250 MB     Downloading: 250 MB     Downloading: 250 MB     Downloading: 250 MB     Downloading: 250 MB     Downloading: 250 MB     Downloading: 250 MB     Downloading: 250 MB     Downloading: 260 MB     Downloading: 260 MB     Downloading: 260 MB     Downloading: 260 MB     Downloading: 260 MB     Downloading: 260 MB     Downloading: 260 MB     Downloading: 260 MB     Downloading: 260 MB     Downloading: 260 MB     Downloading: 260 MB     Downloading: 260 MB     Downloading: 270 MB     Downloading: 270 MB     Downloading: 270 MB     Downloading: 270 MB     Downloading: 270 MB     Downloading: 270 MB     Downloading: 270 MB     Downloading: 270 MB     Downloading: 270 MB     Downloading: 270 MB     Downloading: 270 MB     Downloading: 270 MB     Downloading: 280 MB     Downloading: 280 MB     Downloading: 280 MB     Downloading: 280 MB     Downloading: 280 MB     Downloading: 280 MB     Downloading: 280 MB     Downloading: 280 MB     Downloading: 280 MB     Downloading: 280 MB     Downloading: 280 MB     Downloading: 280 MB     Downloading: 290 MB     Downloading: 290 MB     Downloading: 290 MB     Downloading: 290 MB     Downloading: 290 MB     Downloading: 290 MB     Downloading: 290 MB     Downloading: 290 MB     Downloading: 290 MB     Downloading: 290 MB     Downloading: 290 MB     Downloading: 290 MB     Downloading: 300 MB     Downloading: 300 MB     Downloading: 300 MB     Downloading: 300 MB     Downloading: 300 MB     Downloading: 300 MB     Downloading: 300 MB     Downloading: 300 MB     Downloading: 300 MB     Downloading: 300 MB     Downloading: 300 MB     Downloading: 300 MB     Downloading: 310 MB     Downloading: 310 MB     Downloading: 310 MB     Downloading: 310 MB     Downloading: 310 MB     Downloading: 310 MB     Downloading: 310 MB     Downloading: 310 MB     Downloading: 310 MB     Downloading: 310 MB     Downloading: 310 MB     Downloading: 310 MB     Downloading: 320 MB     Downloading: 320 MB     Downloading: 320 MB     Downloading: 320 MB     Downloading: 320 MB     Downloading: 320 MB     Downloading: 320 MB     Downloading: 320 MB     Downloading: 320 MB     Downloading: 320 MB     Downloading: 320 MB     Downloading: 320 MB     Downloading: 330 MB     Downloading: 330 MB     Downloading: 330 MB     Downloading: 330 MB     Downloading: 330 MB     Downloading: 330 MB     Downloading: 330 MB     Downloading: 330 MB     Downloading: 330 MB     Downloading: 330 MB     Downloading: 330 MB     Downloading: 330 MB     Downloading: 340 MB     Downloading: 340 MB     Downloading: 340 MB     Downloading: 340 MB     Downloading: 340 MB     Downloading: 340 MB     Downloading: 340 MB     Downloading: 340 MB     Downloading: 340 MB     Downloading: 340 MB     Downloading: 340 MB     Downloading: 340 MB     Downloading: 350 MB     Downloading: 350 MB     Downloading: 350 MB     Downloading: 350 MB     Downloading: 350 MB     Downloading: 350 MB     Downloading: 350 MB     Downloading: 350 MB     Downloading: 350 MB     Downloading: 350 MB     Downloading: 350 MB     Downloading: 350 MB     Downloading: 360 MB     Downloading: 360 MB     Downloading: 360 MB     Downloading: 360 MB     Downloading: 360 MB     Downloading: 360 MB     Downloading: 360 MB     Downloading: 360 MB     Downloading: 360 MB     Downloading: 360 MB     Downloading: 360 MB     Downloading: 360 MB     Downloading: 370 MB     Downloading: 370 MB     Downloading: 370 MB     Downloading: 370 MB     Downloading: 370 MB     Downloading: 370 MB     Downloading: 370 MB     Downloading: 370 MB     Downloading: 370 MB     Downloading: 370 MB     Downloading: 370 MB     Downloading: 370 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 380 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 390 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 400 MB     Downloading: 410 MB     Downloading: 410 MB     Downloading: 410 MB     Downloading: 410 MB     Downloading: 410 MB     Downloading: 410 MB     Downloading: 410 MB     Downloading: 410 MB     Downloading: 410 MB     Downloading: 410 MB     Downloading: 410 MB     Downloading: 410 MB     Downloading: 420 MB     Downloading: 420 MB     Downloading: 420 MB     Downloading: 420 MB     Downloading: 420 MB     Downloading: 420 MB     Downloading: 420 MB     Downloading: 420 MB     Downloading: 420 MB     Downloading: 420 MB     Downloading: 420 MB     Downloading: 420 MB     Downloading: 430 MB     Downloading: 430 MB     Downloading: 430 MB     Downloading: 430 MB     Downloading: 430 MB     Downloading: 430 MB     Downloading: 430 MB     Downloading: 430 MB     Downloading: 430 MB     Downloading: 430 MB     Downloading: 430 MB     Downloading: 430 MB     Downloading: 440 MB     Downloading: 440 MB     Downloading: 440 MB     Downloading: 440 MB     Downloading: 440 MB     Downloading: 440 MB     Downloading: 440 MB     Downloading: 440 MB     Downloading: 440 MB     Downloading: 440 MB     Downloading: 440 MB     Downloading: 440 MB     Downloading: 450 MB     Downloading: 450 MB     Downloading: 450 MB     Downloading: 450 MB     Downloading: 450 MB     Downloading: 450 MB     Downloading: 450 MB     Downloading: 450 MB     Downloading: 450 MB     Downloading: 450 MB     Downloading: 450 MB     Downloading: 450 MB     Downloading: 460 MB     Downloading: 460 MB     Downloading: 460 MB     Downloading: 460 MB     Downloading: 460 MB     Downloading: 460 MB     Downloading: 460 MB     Downloading: 460 MB     Downloading: 460 MB     Downloading: 460 MB     Downloading: 460 MB     Downloading: 460 MB     Downloading: 470 MB     Downloading: 470 MB     Downloading: 470 MB     Downloading: 470 MB     Downloading: 470 MB     Downloading: 470 MB     Downloading: 470 MB     Downloading: 470 MB     Downloading: 470 MB     Downloading: 470 MB     Downloading: 470 MB     Downloading: 470 MB     Downloading: 480 MB     Downloading: 480 MB     Downloading: 480 MB     Downloading: 480 MB     Downloading: 480 MB     Downloading: 480 MB     Downloading: 480 MB     Downloading: 480 MB     Downloading: 480 MB     Downloading: 480 MB     Downloading: 480 MB     Downloading: 480 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 490 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 500 MB     Downloading: 510 MB     Downloading: 510 MB     Downloading: 510 MB     Downloading: 510 MB     Downloading: 510 MB     Downloading: 510 MB     Downloading: 510 MB     Downloading: 510 MB     Downloading: 510 MB     Downloading: 510 MB     Downloading: 510 MB     Downloading: 510 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 520 MB     Downloading: 530 MB     Downloading: 530 MB     Downloading: 530 MB     Downloading: 530 MB     Downloading: 530 MB     Downloading: 530 MB     

  ## download demographic data
  zi_get_demographics(year = 2012, variables = "B01003_001", survey = "acs5",
      zcta = mo20$GEOID)
#> # A tibble: 1,024 × 4
#>    GEOID variable   estimate   moe
#>    <chr> <chr>         <dbl> <dbl>
#>  1 51640 B01003_001     1687   145
#>  2 52542 B01003_001      340   126
#>  3 52573 B01003_001       84    61
#>  4 52626 B01003_001     1499   230
#>  5 63005 B01003_001    17741   706
#>  6 63010 B01003_001    36304   985
#>  7 63011 B01003_001    37114   907
#>  8 63012 B01003_001    10334   779
#>  9 63013 B01003_001     1697   489
#> 10 63014 B01003_001      483   150
#> # ℹ 1,014 more rows
# }