Load List of Available Label Data Sets

This function loads a list of available label data sets that can be used to label ZIP Codes. Currently, only three-digit ZIP Codes are supported.

Usage

zi_load_labels_list(type = "zip3")

Arguments

type

A character scalar specifying the type of label data to load. The only supported type is 'zip3' (three-digit ZIP Codes).

Value

A tibble containing date values that can be used with zi_load_labels.

Examples

# \donttest{
  zi_load_labels_list(type = "zip3")
#> # A tibble: 1 × 2
#>   type    date
#>   <chr>  <int>
#> 1 zip3  202408
# }