Missouri HUD ZIP Code to County Crosswalk, 2023

A tibble containing the HUD ZIP Code to County Crosswalk file for Missouri's ZIP Codes in 2023's first quarter.

Usage

data(zi_mo_hud)

Format

A data frame with 1749 rows and 8 variables:

ZIP

five-digit United States Postal Service ZIP Code

GEOID

five-digit county FIPS code

RES_RATIO

for ZIP Codes that cross county boundaries, the proportion of the ZIP Code's residential customers in the given county

BUS_RATIO

for ZIP Codes that cross county boundaries, the proportion of the ZIP Code's commercial customers in the given county

OTH_RATIO

for ZIP Codes that cross county boundaries, the proportion of the ZIP Code's other customers in the given county

TOT_RATIO

for ZIP Codes that cross county boundaries, the proportion of the ZIP Code's total customers in the given county

CITY

United States Postal Service city name

STATE

United States Postal Service state abbreviation

Source

U.S. Department of Housing and Urban Development's ZIP Code crosswalk files

Details

The data included in zi_mo_hud can be replicated with the following code: zi_load_crosswalk(zip_source = "HUD", year = 2023, qtr = 1, target = "COUNTY", query = "MO"). This assumes your HUD API key is stored in your .Rprofile file as hud_key.

Examples

utils::str(zi_mo_hud)
#> tibble [1,749 × 8] (S3: tbl_df/tbl/data.frame)
#>  $ ZIP      : chr [1:1749] "63103" "63447" "63362" "63134" ...
#>  $ GEOID    : chr [1:1749] "29510" "29111" "29113" "29189" ...
#>  $ RES_RATIO: num [1:1749] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ BUS_RATIO: num [1:1749] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ OTH_RATIO: num [1:1749] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ TOT_RATIO: num [1:1749] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ CITY     : chr [1:1749] "SAINT LOUIS" "LA BELLE" "MOSCOW MILLS" "SAINT LOUIS" ...
#>  $ STATE    : chr [1:1749] "MO" "MO" "MO" "MO" ...
utils::head(zi_mo_hud)
#> # A tibble: 6 × 8
#>   ZIP   GEOID RES_RATIO BUS_RATIO OTH_RATIO TOT_RATIO CITY            STATE
#>   <chr> <chr>     <dbl>     <dbl>     <dbl>     <dbl> <chr>           <chr>
#> 1 63103 29510         1         1         1         1 SAINT LOUIS     MO   
#> 2 63447 29111         1         1         1         1 LA BELLE        MO   
#> 3 63362 29113         1         1         1         1 MOSCOW MILLS    MO   
#> 4 63134 29189         1         1         1         1 SAINT LOUIS     MO   
#> 5 65439 29093         1         1         1         1 BIXBY           MO   
#> 6 65686 29209         1         1         1         1 KIMBERLING CITY MO