Total Population and Median Household Income, Missouri ZCTAs 2022

A tibble containing the total population and median household income estimates from the 2018-2022 5-year U.S. Census Bureau American Communiy Survey estimates for Missouri five-digit ZIP Code Tabulation Areas (ZCTAs).

data(zi_mo_pop)

Format

A data frame with 2664 rows and 4 variables:

GEOID

full GEOID string

variable

variable, either B01003_001 (total population) or B19013_001 (median household income)

estimate

value for associated variable

moe

margin of error for associated variable

Source

U.S. Census Bureau American Community Survey

Details

The data included in zi_mo_pop can be replicated with the following code: zi_get_demographics(year = 2022, variables = c("B01003_001", "B19013_001"), survey = "acs5").

Examples

utils::str(zi_mo_pop)
#> tibble [2,664 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ GEOID   : chr [1:2664] "50101" "50101" "50102" "50102" ...
#>  $ variable: chr [1:2664] "B01003_001" "B19013_001" "B01003_001" "B19013_001" ...
#>  $ estimate: num [1:2664] 68 158750 127 NA 400 ...
#>  $ moe     : num [1:2664] 36 146906 87 NA 109 ...
utils::head(zi_mo_pop)
#> # A tibble: 6 × 4
#>   GEOID variable   estimate    moe
#>   <chr> <chr>         <dbl>  <dbl>
#> 1 50101 B01003_001       68     36
#> 2 50101 B19013_001   158750 146906
#> 3 50102 B01003_001      127     87
#> 4 50102 B19013_001       NA     NA
#> 5 50103 B01003_001      400    109
#> 6 50103 B19013_001    55938  10284