Convert HUD Crosswalk Data to Finalized Crosswalk

The output from zi_load_crosswalk() for HUD data requires additional processing to be used in the zi_crosswalk() function. This function prepares the HUD data for use in joins.

zi_prep_hud(.data, by, return_max = TRUE)

Arguments

.data

The output from zi_load_crosswalk() with HUD data.

by

Character scalar; the column name to use for identifying the best match for a given ZIP Code. This could be either "residential", "commercial", or "total".

return_max

Logical scalar; if TRUE (default), only the county with the highest proportion of the ZIP Code type will be returned. If the ZIP Code straddles two states, two records will be returned. If FALSE, all records for the ZIP Code will be returned. Where a tie exists (i.e. two counties each contain half of all addresses), the county with the lowest GEOID value will be returned.

Value

A tibble that has been further prepared for use as a crosswalk.

Examples

if (FALSE) {

  # access to HUD ZIP Code to County crosswalk for all ZIP Codes in Missouri
  mo <- zi_load_crosswalk(zip_source = "HUD", year = 2023, qtr = 1,
    target = "COUNTY", query = "MO")

  # prep data
  mo <- zi_prep_hud(mo, by, return_max = TRUE)

}